+++ docs/sphinx/examples/cpp/other/builder/vqe_h2_builder.cpp
@@ -57,22 +57,23 @@ void so4(cudaq::kernel_builder<std::vector<double>> &builder, QuakeValue &&q,
 int main() {
 
   // Read in the spin op from file
-  std::vector<double> h2_data{15,
-                              -0.10647701149499994, 0, 4, 0, 0, 1, 0, 2, 0, 3, 0, 
-                              0.0454063328691, 0, 4, 0, 2, 1, 2, 2, 2, 3, 2, 
-                              0.0454063328691, 0, 4, 0, 2, 1, 2, 2, 3, 3, 3, 
-                              0.0454063328691, 0, 4, 0, 3, 1, 3, 2, 2, 3, 2, 
-                              0.0454063328691, 0, 4, 0, 3, 1, 3, 2, 3, 3, 3, 
-                              0.170280101353, 0, 4, 0, 1, 1, 0, 2, 0, 3, 0, 
-                              0.120200490713, 0, 4, 0, 1, 1, 1, 2, 0, 3, 0, 
-                              0.168335986252, 0, 4, 0, 1, 1, 0, 2, 1, 3, 0, 
-                              0.165606823582, 0, 4, 0, 1, 1, 0, 2, 0, 3, 1, 
-                              -0.22004130022499996, 0, 4, 0, 0, 1, 1, 2, 0, 3, 0, 
-                              0.165606823582, 0, 4, 0, 0, 1, 1, 2, 1, 3, 0, 
-                              0.174072892497, 0, 4, 0, 0, 1, 1, 2, 0, 3, 1, 
-                              0.170280101353, 0, 4, 0, 0, 1, 0, 2, 1, 3, 0, 
-                              0.120200490713, 0, 4, 0, 0, 1, 0, 2, 1, 3, 1, 
-                              -0.22004130022499996, 0, 4, 0, 0, 1, 0, 2, 0, 3, 1};
+  std::vector<double> h2_data{
+      15, -0.10647701149499994, 0, 4, 0, 0, 1, 0, 2, 0, 3,
+      0,  0.0454063328691,      0, 4, 0, 2, 1, 2, 2, 2, 3,
+      2,  0.0454063328691,      0, 4, 0, 2, 1, 2, 2, 3, 3,
+      3,  0.0454063328691,      0, 4, 0, 3, 1, 3, 2, 2, 3,
+      2,  0.0454063328691,      0, 4, 0, 3, 1, 3, 2, 3, 3,
+      3,  0.170280101353,       0, 4, 0, 1, 1, 0, 2, 0, 3,
+      0,  0.120200490713,       0, 4, 0, 1, 1, 1, 2, 0, 3,
+      0,  0.168335986252,       0, 4, 0, 1, 1, 0, 2, 1, 3,
+      0,  0.165606823582,       0, 4, 0, 1, 1, 0, 2, 0, 3,
+      1,  -0.22004130022499996, 0, 4, 0, 0, 1, 1, 2, 0, 3,
+      0,  0.165606823582,       0, 4, 0, 0, 1, 1, 2, 1, 3,
+      0,  0.174072892497,       0, 4, 0, 0, 1, 1, 2, 0, 3,
+      1,  0.170280101353,       0, 4, 0, 0, 1, 0, 2, 1, 3,
+      0,  0.120200490713,       0, 4, 0, 0, 1, 0, 2, 1, 3,
+      1,  -0.22004130022499996, 0, 4, 0, 0, 1, 0, 2, 0, 3,
+      1};
   cudaq::spin_op H(h2_data);
 
   int layers = 2, n_qubits = H.num_qubits(), block_size = 2, p_counter = 0;
