+++ docs/sphinx/examples/cpp/other/compute_actions.cpp
@@ -67,22 +67,23 @@ struct ansatz_compute_action {
 
 int main(int argc, char **argv) {
 
-  std::vector<double> h2_data{15,
-                              -0.10647701149499994, 0, 4, 0, 0, 1, 0, 2, 0, 3, 0, 
-                              0.0454063328691, 0, 4, 0, 2, 1, 2, 2, 2, 3, 2, 
-                              0.0454063328691, 0, 4, 0, 2, 1, 2, 2, 3, 3, 3, 
-                              0.0454063328691, 0, 4, 0, 3, 1, 3, 2, 2, 3, 2, 
-                              0.0454063328691, 0, 4, 0, 3, 1, 3, 2, 3, 3, 3, 
-                              0.170280101353, 0, 4, 0, 1, 1, 0, 2, 0, 3, 0, 
-                              0.120200490713, 0, 4, 0, 1, 1, 1, 2, 0, 3, 0, 
-                              0.168335986252, 0, 4, 0, 1, 1, 0, 2, 1, 3, 0, 
-                              0.165606823582, 0, 4, 0, 1, 1, 0, 2, 0, 3, 1, 
-                              -0.22004130022499996, 0, 4, 0, 0, 1, 1, 2, 0, 3, 0, 
-                              0.165606823582, 0, 4, 0, 0, 1, 1, 2, 1, 3, 0, 
-                              0.174072892497, 0, 4, 0, 0, 1, 1, 2, 0, 3, 1, 
-                              0.170280101353, 0, 4, 0, 0, 1, 0, 2, 1, 3, 0, 
-                              0.120200490713, 0, 4, 0, 0, 1, 0, 2, 1, 3, 1, 
-                              -0.22004130022499996, 0, 4, 0, 0, 1, 0, 2, 0, 3, 1};
+  std::vector<double> h2_data{
+      15, -0.10647701149499994, 0, 4, 0, 0, 1, 0, 2, 0, 3,
+      0,  0.0454063328691,      0, 4, 0, 2, 1, 2, 2, 2, 3,
+      2,  0.0454063328691,      0, 4, 0, 2, 1, 2, 2, 3, 3,
+      3,  0.0454063328691,      0, 4, 0, 3, 1, 3, 2, 2, 3,
+      2,  0.0454063328691,      0, 4, 0, 3, 1, 3, 2, 3, 3,
+      3,  0.170280101353,       0, 4, 0, 1, 1, 0, 2, 0, 3,
+      0,  0.120200490713,       0, 4, 0, 1, 1, 1, 2, 0, 3,
+      0,  0.168335986252,       0, 4, 0, 1, 1, 0, 2, 1, 3,
+      0,  0.165606823582,       0, 4, 0, 1, 1, 0, 2, 0, 3,
+      1,  -0.22004130022499996, 0, 4, 0, 0, 1, 1, 2, 0, 3,
+      0,  0.165606823582,       0, 4, 0, 0, 1, 1, 2, 1, 3,
+      0,  0.174072892497,       0, 4, 0, 0, 1, 1, 2, 0, 3,
+      1,  0.170280101353,       0, 4, 0, 0, 1, 0, 2, 1, 3,
+      0,  0.120200490713,       0, 4, 0, 0, 1, 0, 2, 1, 3,
+      1,  -0.22004130022499996, 0, 4, 0, 0, 1, 0, 2, 0, 3,
+      1};
   cudaq::spin_op H(h2_data);
   const auto param_space = cudaq::linspace(-M_PI, M_PI, 10);
   printf("Using the hand-coded kernel\n");
