+++ docs/sphinx/examples/cpp/other/distributed/mpi.cpp
@@ -29,9 +29,10 @@ int main(int argc, char **argv) {
   if (cudaq::mpi::rank() == 0)
     printf("Running MPI example with %d processes.\n", cudaq::mpi::num_ranks());
 
-  cudaq::spin_op h = 5.907 - 2.1433 * cudaq::spin_op::x(0) * cudaq::spin_op::x(1) - 
-                     2.1433 * cudaq::spin_op::y(0) * cudaq::spin_op::y(1) +
-                     .21829 * cudaq::spin_op::z(0) - 6.125 * cudaq::spin_op::z(1);
+  cudaq::spin_op h =
+      5.907 - 2.1433 * cudaq::spin_op::x(0) * cudaq::spin_op::x(1) -
+      2.1433 * cudaq::spin_op::y(0) * cudaq::spin_op::y(1) +
+      .21829 * cudaq::spin_op::z(0) - 6.125 * cudaq::spin_op::z(1);
   auto ansatz = [](double theta) __qpu__ {
     cudaq::qubit q, r;
     x(q);
