+++ docs/sphinx/examples/cpp/other/remote_sim/simple.cpp
@@ -22,9 +22,10 @@
 // its features and command line options.
 
 int main() {
-  cudaq::spin_op h = 5.907 - 2.1433 * cudaq::spin_op::x(0) * cudaq::spin_op::x(1) - 
-                     2.1433 * cudaq::spin_op::y(0) * cudaq::spin_op::y(1) +
-                     .21829 * cudaq::spin_op::z(0) - 6.125 * cudaq::spin_op::z(1);
+  cudaq::spin_op h =
+      5.907 - 2.1433 * cudaq::spin_op::x(0) * cudaq::spin_op::x(1) -
+      2.1433 * cudaq::spin_op::y(0) * cudaq::spin_op::y(1) +
+      .21829 * cudaq::spin_op::z(0) - 6.125 * cudaq::spin_op::z(1);
 
   auto [ansatz, theta] = cudaq::make_kernel<double>();
   // Allocate some qubits
