+++ docs/sphinx/snippets/cpp/using/cudaq/platform/observe_mqpu_mpi.cpp
@@ -16,9 +16,10 @@
 int main() {
   // [Begin Documentation]
   cudaq::mpi::initialize();
-  cudaq::spin_op h = 5.907 - 2.1433 * cudaq::spin_op::x(0) * cudaq::spin_op::x(1) - 
-                     2.1433 * cudaq::spin_op::y(0) * cudaq::spin_op::y(1) +
-                     .21829 * cudaq::spin_op::z(0) - 6.125 * cudaq::spin_op::z(1);
+  cudaq::spin_op h =
+      5.907 - 2.1433 * cudaq::spin_op::x(0) * cudaq::spin_op::x(1) -
+      2.1433 * cudaq::spin_op::y(0) * cudaq::spin_op::y(1) +
+      .21829 * cudaq::spin_op::z(0) - 6.125 * cudaq::spin_op::z(1);
 
   auto ansatz = [](double theta) __qpu__ {
     cudaq::qubit q, r;
