+++ runtime/cudaq/dynamics/CMakeLists.txt
@@ -12,13 +12,13 @@ set(INTERFACE_POSITION_INDEPENDENT_CODE ON)
 
 set(CUDAQ_OPS_SRC
   callback.cpp
-  scalar_operators.cpp
-  spin_operators.cpp
-  boson_operators.cpp
-  fermion_operators.cpp
-  matrix_operators.cpp
-  product_operators.cpp
-  operator_sum.cpp
+  scalar_op.cpp
+  spin_op.cpp
+  boson_op.cpp
+  fermion_op.cpp
+  matrix_op.cpp
+  product_op.cpp
+  sum_op.cpp
   rydberg_hamiltonian.cpp
   evolution.cpp
   handler.cpp
