+++ runtime/cudaq/dynamics/matrix_op.cpp
@@ -13,10 +13,10 @@
 #include "cudaq/operators.h"
 #include "cudaq/utils/matrix.h"
 
-#include "boson_operators.h"
-#include "fermion_operators.h"
-#include "matrix_operators.h"
-#include "spin_operators.h"
+#include "cudaq/boson_op.h"
+#include "cudaq/fermion_op.h"
+#include "cudaq/matrix_op.h"
+#include "cudaq/spin_op.h"
 
 namespace cudaq {
 
@@ -506,12 +506,24 @@ product_op<matrix_handler> matrix_handler::squeeze(std::size_t degree) {
 }
 
 namespace operators {
-product_op<matrix_handler> number(std::size_t target) { return matrix_handler::number(target); }
-product_op<matrix_handler> parity(std::size_t target) { return matrix_handler::parity(target); }
-product_op<matrix_handler> position(std::size_t target) { return matrix_handler::position(target); }
-product_op<matrix_handler> momentum(std::size_t target) { return matrix_handler::momentum(target); }
-product_op<matrix_handler> squeeze(std::size_t target) { return matrix_handler::squeeze(target); }
-product_op<matrix_handler> displace(std::size_t target) { return matrix_handler::displace(target); }
-} // namespace boson
+product_op<matrix_handler> number(std::size_t target) {
+  return matrix_handler::number(target);
+}
+product_op<matrix_handler> parity(std::size_t target) {
+  return matrix_handler::parity(target);
+}
+product_op<matrix_handler> position(std::size_t target) {
+  return matrix_handler::position(target);
+}
+product_op<matrix_handler> momentum(std::size_t target) {
+  return matrix_handler::momentum(target);
+}
+product_op<matrix_handler> squeeze(std::size_t target) {
+  return matrix_handler::squeeze(target);
+}
+product_op<matrix_handler> displace(std::size_t target) {
+  return matrix_handler::displace(target);
+}
+} // namespace operators
 
 } // namespace cudaq
