+++ scripts/build_cudaq.sh
@@ -74,6 +74,9 @@ repo_root=$(cd "$this_file_dir" && git rev-parse --show-toplevel)
 # Prepare the build directory
 build_dir="$working_dir/build"
 echo "Build directory: $build_dir"
+if [ "${build_configuration,,}" == "release" ]; then
+  build_dir+="/release"
+fi
 mkdir -p "$CUDAQ_INSTALL_PREFIX/bin"
 mkdir -p "$build_dir" && cd "$build_dir" # && rm -rf * 
 mkdir -p logs && rm -rf logs/*
